SOCINT-329 move RefusalType
